🔨 Service command와 query 분리, Q 클래스 static import
