+++ src/main/java/qastudio/backend/domain/project/repository/PageRole/PageRoleRepositoryImpl.java
@@ -4,8 +4,8 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import qastudio.backend.domain.project.entity.PageRole;
-import qastudio.backend.domain.project.entity.QPage;
-import qastudio.backend.domain.project.entity.QPageRole;
+import static qastudio.backend.domain.project.entity.QPage.page;
+import static qastudio.backend.domain.project.entity.QPageRole.pageRole;
 
 import java.util.List;
 
@@ -13,14 +13,12 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PageRoleRepositoryImpl implements PageRoleRepositoryCustom {
     private final JPAQueryFactory jpaQueryFactory;
-    private final QPage qPage = QPage.page;
-    private final QPageRole qPageRole = QPageRole.pageRole;
 
     @Override
     public List<PageRole> findAllByPageId(Long pageId) {
         return jpaQueryFactory
-                .selectFrom(qPageRole)
-                .where(qPageRole.page.id.eq(pageId))
+                .selectFrom(pageRole)
+                .where(pageRole.page.id.eq(pageId))
                 .fetch();
     }
 }
