+++ src/main/java/qastudio/backend/domain/project/repository/PageSceenario/PageScenarioRepositoryImpl.java
@@ -4,8 +4,8 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import qastudio.backend.domain.project.entity.PageScenario;
-import qastudio.backend.domain.project.entity.QPage;
-import qastudio.backend.domain.project.entity.QPageScenario;
+import static qastudio.backend.domain.project.entity.QPage.page;
+import static qastudio.backend.domain.project.entity.QPageScenario.pageScenario;
 
 import java.util.List;
 
@@ -13,14 +13,12 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PageScenarioRepositoryImpl implements PageScenarioRepositoryCustom {
     private final JPAQueryFactory jpaQueryFactory;
-    private final QPage qPage = QPage.page;
-    private final QPageScenario qPageScenario = QPageScenario.pageScenario;
 
     @Override
     public List<PageScenario> findAllByPageId(Long pageId) {
         return jpaQueryFactory
-                .selectFrom(qPageScenario)
-                .where(qPageScenario.page.id.eq(pageId))
+                .selectFrom(pageScenario)
+                .where(pageScenario.page.id.eq(pageId))
                 .fetch();
     }
 }
