+++ src/main/java/qastudio/backend/domain/project/service/PageCommandService.java
@@ -1,0 +1,10 @@
+package qastudio.backend.domain.project.service;
+
+import qastudio.backend.domain.project.dto.request.PageRequest;
+import qastudio.backend.domain.project.dto.response.PageResponse;
+
+public interface PageCommandService {
+    PageResponse.PageSummary createPage(Long projectId, PageRequest.createPage createPage);
+
+    void deletePage(Long pageId);
+}
