+++ src/main/java/qastudio/backend/domain/project/service/PageQueryService.java
@@ -8,9 +8,5 @@ import qastudio.backend.domain.project.entity.Page;
 import java.util.List;
 
 public interface PageQueryService {
-    PageResponse.PageSummary createPage(Long projectId, PageRequest.createPage createPage);
-
-    void deletePage(Long pageId);
-
     List<PageResponse.PageSummary> getAllPage(Long projectId);
 }
